refactor(Profile, ProfileController): move profile to project domain and update endpoints
