+++ apps/backend/src/main/java/scrumpledpaper/agiler/note/entity/NoteProfile.java
@@ -10,7 +10,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import scrumpledpaper.agiler.user.entity.Profile;
+import scrumpledpaper.agiler.project.entity.Profile;
 import scrumpledpaper.agiler.common.BaseEntity;
 
 @Getter
