+++ apps/backend/src/main/java/scrumpledpaper/agiler/project/entity/Project.java
@@ -1,5 +1,8 @@
 package scrumpledpaper.agiler.project.entity;
 
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,6 +14,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import scrumpledpaper.agiler.common.BaseEntity;
+import scrumpledpaper.agiler.project.dto.ProfileResDto;
+import scrumpledpaper.agiler.user.entity.User;
 
 @Getter
 @NoArgsConstructor
@@ -41,4 +46,6 @@ public class Project extends BaseEntity {
 		this.url = url;
 		this.summary = summary;
 	}
+
+
 }
