+++ apps/backend/src/main/java/scrumpledpaper/agiler/project/entity/Role.java
@@ -1,0 +1,6 @@
+package scrumpledpaper.agiler.project.entity;
+
+public enum Role {
+	MEMBER,
+	OWNER
+}
