+++ apps/backend/src/main/java/scrumpledpaper/agiler/project/mapper/ProfileMapper.java
@@ -1,13 +1,12 @@
-package scrumpledpaper.agiler.user.mapper;
+package scrumpledpaper.agiler.project.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
 import scrumpledpaper.agiler.project.dto.ProfileResDto;
-import scrumpledpaper.agiler.project.dto.ProjectSideResDto;
+import scrumpledpaper.agiler.project.entity.Profile;
 import scrumpledpaper.agiler.project.entity.Project;
-import scrumpledpaper.agiler.user.entity.Profile;
-import scrumpledpaper.agiler.user.entity.Role;
+import scrumpledpaper.agiler.project.entity.Role;
 import scrumpledpaper.agiler.user.entity.User;
 
 @Mapper(componentModel = "spring")
