+++ apps/backend/src/main/java/scrumpledpaper/agiler/project/repository/ProfileRepository.java
@@ -1,14 +1,12 @@
-package scrumpledpaper.agiler.user.repository;
+package scrumpledpaper.agiler.project.repository;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-import scrumpledpaper.agiler.user.entity.Profile;
+import scrumpledpaper.agiler.project.entity.Profile;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
 	Optional<Profile> findByUserIdAndProjectId(Long userId, Long projectId);
