+++ apps/backend/src/main/java/scrumpledpaper/agiler/project/service/ProfileService.java
@@ -1,4 +1,4 @@
-package scrumpledpaper.agiler.user.service;
+package scrumpledpaper.agiler.project.service;
 
 import java.util.Optional;
 
@@ -11,20 +11,18 @@ import scrumpledpaper.agiler.common.exception.CustomException;
 import scrumpledpaper.agiler.common.exception.ErrorCode;
 import scrumpledpaper.agiler.image.service.ImageService;
 import scrumpledpaper.agiler.project.dto.ProfileResDto;
+import scrumpledpaper.agiler.project.entity.Profile;
 import scrumpledpaper.agiler.project.entity.Project;
-import scrumpledpaper.agiler.project.repository.ProjectRepository;
-import scrumpledpaper.agiler.user.entity.Profile;
-import scrumpledpaper.agiler.user.entity.Role;
+import scrumpledpaper.agiler.project.entity.Role;
+import scrumpledpaper.agiler.project.mapper.ProfileMapper;
+import scrumpledpaper.agiler.project.repository.ProfileRepository;
 import scrumpledpaper.agiler.user.entity.User;
-import scrumpledpaper.agiler.user.mapper.ProfileMapper;
-import scrumpledpaper.agiler.user.repository.ProfileRepository;
 
 @Service
 @RequiredArgsConstructor
 public class ProfileService {
 	private final ProfileMapper profileMapper;
 	private final ImageService imageService;
-	private final ProjectRepository projectRepository;
 	private final ProfileRepository profileRepository;
 
 	public void createDefaultProfile(User user, Project savedProject, Role role) {
@@ -54,17 +52,5 @@ public class ProfileService {
 
 				return profileMapper.toProfileResDto(profile, imageUrl);
 			});
-	}
-
-	public ProfileResDto getProjectProfile(Long id, String ProjectUrl) {
-		Project project = projectRepository.findByUrl(ProjectUrl)
-			.orElseThrow(() -> new CustomException(ErrorCode.PROJECT_NOT_FOUND));
-		Profile profile = getProfileByUserIdAndProjectId(id, project.getId());
-
-		String imageUrl = Optional.ofNullable(profile.getUser().getImageId())
-			.map(imageService::getImageUrlById)
-			.orElse("");
-
-		return profileMapper.toProfileResDto(profile, imageUrl);
 	}
 }
