--- apps/backend/src/main/java/scrumpledpaper/agiler/user/entity/Role.java
@@ -1,6 +1,0 @@
-package scrumpledpaper.agiler.user.entity;
-
-public enum Role {
-	MEMBER,
-	OWNER
-}
