+++ apps/backend/src/test/java/scrumpledpaper/agiler/user/ProfileControllerTest.java
@@ -13,17 +13,19 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import scrumpledpaper.agiler.annotation.IntegrationTest;
 import scrumpledpaper.agiler.common.AuthContext;
+import scrumpledpaper.agiler.common.PageResDto;
 import scrumpledpaper.agiler.common.TestDataFactory;
 import scrumpledpaper.agiler.common.exception.ErrorCode;
 import scrumpledpaper.agiler.image.entity.Image;
 import scrumpledpaper.agiler.project.dto.ProfileResDto;
+import scrumpledpaper.agiler.project.entity.Profile;
 import scrumpledpaper.agiler.project.entity.Project;
-import scrumpledpaper.agiler.user.entity.Profile;
-import scrumpledpaper.agiler.user.entity.Role;
+import scrumpledpaper.agiler.project.entity.Role;
 
 @IntegrationTest
 @Transactional
@@ -36,6 +38,109 @@ public class ProfileControllerTest {
 	private TestDataFactory testDataFactory;
 	Image defaultImage;
 
+
+	@Nested
+	@DisplayName("Get Project Member List Test")
+	class GetProjectMemberTest {
+		@BeforeEach
+		void beforeEach() {
+			defaultImage = testDataFactory.createDefaultImage();
+		}
+
+		@Test
+		@DisplayName("200 - 프로젝트 멤버 조회 성공")
+		void getProjectProfilesSuccess() throws Exception {
+			// given
+			AuthContext auth = testDataFactory.createAuth(defaultImage);
+			String url = "test_url";
+			Project project = testDataFactory.createProjectAndOwnerProfile(url, auth.getUser());
+			AuthContext member1 = testDataFactory.createAuth(testDataFactory.createDefaultImage());
+			AuthContext member2 = testDataFactory.createAuth(testDataFactory.createDefaultImage());
+			testDataFactory.createProfile(member1.getUser(), project, Role.MEMBER);
+			testDataFactory.createProfile(member2.getUser(), project, Role.MEMBER);
+
+			// when
+			String response = mockMvc.perform(
+					get("/api/v1/projects/{projectUrl}/profiles", url)
+						.header("Authorization", auth.bearer())
+						.param("page", "0")
+						.param("size", "10"))
+				.andExpect(status().isOk())
+				.andReturn().getResponse().getContentAsString();
+
+			// then
+			PageResDto<ProfileResDto> page = objectMapper.readValue(response,
+				new TypeReference<PageResDto<ProfileResDto>>() {
+				});
+
+			assertThat(page.getTotalElements()).isEqualTo(3);
+			assertThat(page.getContents()).hasSize(3);
+		}
+
+		@Test
+		@DisplayName("404 - 존재하지 않는 프로젝트 조회 실패")
+		void getProjectProfilesNotFound() throws Exception {
+			// given
+			AuthContext auth = testDataFactory.createAuth(defaultImage);
+			String nonExistingUrl = "non-existing_url";
+
+			// when
+			String response = mockMvc.perform(
+					get("/api/v1/projects/{projectUrl}/profiles", nonExistingUrl)
+						.header("Authorization", auth.bearer())
+						.param("page", "0")
+						.param("size", "10"))
+				.andExpect(status().isNotFound())
+				.andReturn().getResponse().getContentAsString();
+
+			// then
+			assertThat(response).contains(ErrorCode.PROJECT_NOT_FOUND.getCode());
+		}
+
+		@Test
+		@DisplayName("403 - 프로젝트 멤버가 아닌 유저의 조회 실패")
+		void getProjectProfilesForbidden() throws Exception {
+			// given
+			AuthContext auth = testDataFactory.createAuth(defaultImage);
+			AuthContext anotherAuth = testDataFactory.createAuth(defaultImage);
+			String url = "test_url";
+			Project project = testDataFactory.createProjectAndOwnerProfile(url, auth.getUser());
+
+			// when
+			String response = mockMvc.perform(
+					get("/api/v1/projects/{projectUrl}/profiles", url)
+						.header("Authorization", anotherAuth.bearer())
+						.param("page", "0")
+						.param("size", "10"))
+				.andExpect(status().isForbidden())
+				.andReturn().getResponse().getContentAsString();
+
+			// then
+			assertThat(response).contains(ErrorCode.PROJECT_NOT_MEMBER.getCode());
+		}
+
+		@Test
+		@DisplayName("400 - 잘못된 페이지 요청")
+		void getProjectProfilesInvalidPage() throws Exception {
+			// given
+			AuthContext auth = testDataFactory.createAuth(defaultImage);
+			String url = "test_url";
+			Project project = testDataFactory.createProjectAndOwnerProfile(url, auth.getUser());
+
+			// when
+			String response = mockMvc.perform(
+					get("/api/v1/projects/{projectUrl}/profiles", url)
+						.header("Authorization", auth.bearer())
+						.param("page", "99")
+						.param("size", "10"))
+				.andExpect(status().isNotFound())
+				.andReturn().getResponse().getContentAsString();
+
+			// then
+			assertThat(response).contains(ErrorCode.PAGE_NOT_FOUND.getCode());
+		}
+	}
+
 	@Nested
 	@DisplayName("프로필 조회")
 	class GetProfile {
@@ -45,8 +150,8 @@ public class ProfileControllerTest {
 		}
 
 		@Test
-		@DisplayName("200 - 프로젝트 프로필 조회 성공")
-		public void getProjectProfileOwnerSuccess() throws Exception {
+		@DisplayName("200 - 내 프로젝트 프로필 조회 성공")
+		public void getMyProjectProfileOwnerSuccess() throws Exception {
 			// given
 			AuthContext auth = testDataFactory.createAuth(defaultImage);
 			String url = "test_url";
@@ -55,7 +160,7 @@ public class ProfileControllerTest {
 
 			// when
 			String response = mockMvc.perform(
-					get("/api/v1/profiles/{projectUrl}", url)
+					get("/api/v1/project/{projectUrl}/me", url)
 						.header("Authorization", auth.bearer())
 						.contentType(MediaType.APPLICATION_JSON))
 				.andExpect(status().isOk())
@@ -73,7 +178,7 @@ public class ProfileControllerTest {
 
 		@Test
 		@DisplayName("200 - 프로젝트 프로필 조회 성공 - member")
-		public void getProjectProfileMemberSuccess() throws Exception {
+		public void getMyProjectProfileMemberSuccess() throws Exception {
 			// given
 			AuthContext auth = testDataFactory.createAuth(defaultImage);
 			AuthContext memberAuth = testDataFactory.createAuth(defaultImage);
@@ -83,7 +188,7 @@ public class ProfileControllerTest {
 
 			// when
 			String response = mockMvc.perform(
-					get("/api/v1/profiles/{projectUrl}", url)
+					get("/api/v1/project/{projectUrl}/me", url)
 						.header("Authorization", memberAuth.bearer())
 						.contentType(MediaType.APPLICATION_JSON))
 				.andExpect(status().isOk())
