+++ apps/backend/src/testFixtures/java/scrumpledpaper/agiler/fixture/ProfileFixture.java
@@ -1,8 +1,8 @@
 package scrumpledpaper.agiler.fixture;
 
 import scrumpledpaper.agiler.project.entity.Project;
-import scrumpledpaper.agiler.user.entity.Profile;
-import scrumpledpaper.agiler.user.entity.Role;
+import scrumpledpaper.agiler.project.entity.Profile;
+import scrumpledpaper.agiler.project.entity.Role;
 import scrumpledpaper.agiler.user.entity.User;
 
 public class ProfileFixture {
